Coding Days
git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/trunk@2679 4eea3252-f0fb-4393-894d-40516dce545b

+++ src/sgpp/base/grid/GridDataBase.cpp
@@ -22,9 +22,9 @@ namespace sg
 namespace base
 {
 
-  GridDataBase::GridDataBase(size_t dim) : _map(), _dim(dim) {}
+  GridDataBase::GridDataBase(size_t dim) : _map(), _dim(static_cast<int>(dim)) {}
 
-  GridDataBase::GridDataBase(Grid *grid, DataVector &values) : _map(), _dim(grid->getStorage()->dim()) {
+  GridDataBase::GridDataBase(Grid *grid, DataVector &values) : _map(), _dim(static_cast<int>(grid->getStorage()->dim())) {
     GridStorage* gs = grid->getStorage();
     for (size_t i = 0; i < gs->size(); i++) {
       set(gs->get(i), values[gs->seq(gs->get(i))]);
